renamed module to r_package...easier to find?
