+++ lib/ansible/modules/packaging/language/r_package.py
@@ -3,9 +3,6 @@
 # Copyright: (c) 2018, Evan Van Dam <evandam92@gmail.com>
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 
-from __future__ import absolute_import, division, print_function
-__metaclass__ = type
-
 ANSIBLE_METADATA = {
     'metadata_version': '1.1',
     'status': ['preview'],
@@ -20,7 +17,7 @@ description:
     Install, remove, and update R packages.
     Flexibility is provided to specify the R executable,
     the remote repository, source to build, and library location.
-version_added: "2.8"
+version_added: "2.7"
 author: "Evan Van Dam (@evandam)"
 options:
   name:
@@ -29,7 +26,7 @@ options:
   state:
     description: The target state of the package.
     choices: ["present", "absent"]
-    default: present
+    required: true
   force:
     description: Force the package to be installed, even if it already exists.
     type: bool
@@ -38,15 +35,11 @@ options:
     description: Path to the R executable. If not specified, find it from the PATH.
   src:
     description: The source file or directory to build the R package.
-  type:
-    description: Type of package to install, like "source" or "binary".
   lib:
     description: The library location to use to install or remove the package.
   repo:
     description: The CRAN repo to use to install the package.
-notes:
-  - Requires R to already be installed.
-  - check mode is supported.
+notes: Requires R to already be installed. check_mode is supported.
 '''
 
 EXAMPLES = '''
@@ -71,15 +64,11 @@ EXAMPLES = '''
   r_package:
     name: h2o
     src: ~/Downloads/h2o.tar.gz
-    type: source
 '''
 
 RETURN = '''
-cmd:
-  description: The R function executed by the module.
-  returned: changed
-  type: string
-  sample: install.packages("dplyr")
+msg:
+  description: The output from the R command to install/remove the package.
 '''
 
 import os
@@ -89,7 +78,7 @@ from ansible.module_utils.basic import AnsibleModule
 
 def is_package_installed(module, exe, name, lib=None):
     """Check if the package is installed.
-
+    
     :param module: The current ansible module
     :param exe: The R executable to run
     :param name: The R package to check
@@ -98,33 +87,34 @@ def is_package_installed(module, exe, name, lib=None):
     """
     # lib.loc isn't a valid Python variable,
     # so pass kwargs as a dict and unpack it.
-    func = build_r_function('find.package', name, **{'lib.loc': lib})
-    retcode, out, err = run_command(module, exe, func, check_rc=False)
+    retcode, _, _ = run_command(module, exe, 'find.package', name,
+                                check_rc=False, **{'lib.loc': lib})
+    # an error is thrown if the package isn't found,
+    # so check if it exited successfully.
     return retcode == 0
 
 
-def install_package(module, exe, name, src=None, pkg_type=None, lib=None, repo=None):
+def install_package(module, exe, name, src=None, lib=None, repo=None):
     """Attempt to install the R package.
-
+    
     :param module: The current ansible module
     :param exe: The R executable to run
     :param name: The name of the R package to install
     :param src: The source file/directory of the R package to install. Optional.
-    :param pkg_type: The package type. Ex: "source", "binary", "mac.binary". Optional.
     :param lib: The library to use to install the package. Optional.
     :param repo: The CRAN mirror/repo to use. Optional.
-    :return: The function string that was executed.
+    :return: The output of the `install.packages` R command.
     """
     # Install the package from the repo to the lib.
     install_name = src or name
-    func = build_r_function('install.packages', install_name, type=pkg_type, lib=lib, repos=repo)
-    run_command(module, exe, func, check_rc=True)
-    return func
+    _, out, _ = run_command(module, exe, 'install.packages', install_name,
+                            lib=lib, repos=repo, check_rc=True)
+    return out
 
 
 def remove_package(module, exe, name, lib=None):
     """Attempt to remove the R package.
-
+    
     :param module: The current ansible module
     :param exe: The R executable to run
     :param name: The name of the R package to install
@@ -132,9 +122,9 @@ def remove_package(module, exe, name, lib=None):
     :return: The output of the `remove.packages` R command.
     """
     # Remove the package from the specified lib or the default location.
-    func = build_r_function('remove.packages', name, lib=lib)
-    module.run_command(module, exe, func, lib=lib, check_rc=True)
-    return func
+    _, out, _ = run_command(module, exe, 'remove.packages', name,
+                            lib=lib, check_rc=True)
+    return out
 
 
 def _find_r(module):
@@ -143,7 +133,7 @@ def _find_r(module):
     if executable:
         if os.path.isfile(executable):
             return executable
-        module.fail_json(msg='{0} cannot be found or is not a file'.format(executable))
+        module.fail_json(msg='{} cannot be found or is not a file'.format(executable))
     return module.get_bin_path('R', required=True)
 
 
@@ -151,9 +141,7 @@ def _escape_arg(arg):
     """Strings must be quoted. Numbers are fine as is."""
     if isinstance(arg, Number):
         return arg
-    elif arg is None:
-        return 'NULL'
-    return '"{0}"'.format(arg)
+    return '"{}"'.format(arg)
 
 
 def build_r_function(funcname, *args, **kwargs):
@@ -169,15 +157,24 @@ def build_r_function(funcname, *args, **kwargs):
     return funcname + '(' + ', '.join(r_args) + ')'
 
 
-def run_command(module, exe, func, check_rc=False):
+def run_command(module, exe, funcname, *args, **kwargs):
     """Build the R function and run the command.
 
     :param module: The current ansible module
     :param exe: The R executable to run
-    :param func: The full R function to run.
-    :param check_rc: module.fail_json() if the command fails.
+    :param funcname: The name of the R function to run.
+    :param args: Positional arguments to pass to the R function.
+    :param kwargs: Keyword arguments to pass to the R function.
     :return: A tuple of the returncode, stdout, and stderr of the command.
     """
+    # Intercept the check_rc arg and use it to control module.run_command.
+    try:
+        check_rc = kwargs.pop('check_rc')
+    except KeyError:
+        check_rc = False
+
+    # Build the function string to call in R
+    func = build_r_function(funcname, *args, **kwargs)
     cmd = [exe, '--slave', '--no-save', '--no-restore', '-e', func]
     return module.run_command(cmd, check_rc=check_rc)
 
@@ -191,7 +188,6 @@ def main():
             force=dict(default=False, type='bool'),
             executable=dict(required=False),
             src=dict(required=False),
-            type=dict(required=False),
             lib=dict(required=False),
             repo=dict(required=False),
         ),
@@ -202,23 +198,22 @@ def main():
     name = module.params['name']
     src = module.params['src']
     state = module.params['state']
-    pkg_type = module.params['type']
     force = module.params['force']
     lib = module.params['lib']
     repo = module.params['repo']
 
     installed = is_package_installed(module, exe, name, lib)
     changed = False
-    cmd = ''
+    msg = ''
     if state == 'absent' and installed:
         if not module.check_mode:
-            cmd = remove_package(module, exe, name, lib=lib)
+            msg = remove_package(module, exe, name, lib=lib)
         changed = True
     elif state == 'present' and (not installed or force):
         if not module.check_mode:
-            cmd = install_package(module, exe, name, src, pkg_type, lib=lib, repo=repo)
+            msg = install_package(module, exe, name, src, lib=lib, repo=repo)
         changed = True
-    module.exit_json(changed=changed, cmd=cmd)
+    module.exit_json(changed=changed, msg=msg)
 
 
 if __name__ == '__main__':
